@@ -11,7 +11,8 @@
     "start": "node ./server/server.js",
     "build": "node ./node_modules/webpack/bin/webpack.js -w",
     "seed": "node ./server/seed.js seed",
-    "seedDebug": "node %NODE_DEBUG_OPTION% ./server/seed.js seed"
+    "seedDebug": "node %NODE_DEBUG_OPTION% ./server/seed.js seed",
+    "postinstall": "patch-package"
   },
   "author": "",
   "license": "ISC",
@@ -90,15 +91,17 @@
     "passport-yahoo": "^0.3.0",
     "passport-yahoo-oauth2": "https://github.com/francisalvinbarretto/passport-yahoo-oauth/tarball/master",
     "passport-yahoo-token": "^2.1.0",
+    "patch-package": "^6.0.2",
     "pg": "6.4.1",
     "prelude-extension": "^0.1.0",
     "prelude-ls": "^1.1.2",
-    "react": "^16.4.0",
+    "react": "^16.8.6",
     "react-addons-shallow-compare": "^15.4.1",
     "react-beautiful-dnd": "^3.0.0",
     "react-bootstrap": "^0.32.1",
     "react-bootstrap-date-picker": "^5.1.0",
     "react-bootstrap-typeahead": "^2.6.0",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-currency-formatter": "^1.1.0",
     "react-date-picker-cs": "^0.1.28",
     "react-datepicker": "^1.5.0",
@@ -111,6 +114,7 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.8",
+    "react-selectize": "^3.0.1",
     "react-social-login-buttons": "^1.3.5",
     "react-table": "^6.8.6",
     "react-toggle": "^4.0.2",
@@ -118,6 +122,7 @@
     "react-widgets": "^4.2.6",
     "react-widgets-moment": "^4.0.12",
     "reactstrap": "^6.3.0",
+    "recharts": "^1.5.0",
     "redux": "^3.6.0",
     "redux-auth-wrapper": "^2.0.2",
     "redux-immutable": "^4.0.0",
