@@ -1,9 +1,10 @@
-import React from 'react'
-import {Link} from 'react-router-dom'
+import React      from 'react'
+import {Link}     from 'react-router-dom'
 import ReactTable from 'react-table'
-import Moment from 'moment'
-import csv from 'csv-file-creator'
-import update from "immutability-helper";
+import Moment     from 'moment'
+import 'moment/locale/en-gb'
+import csv        from 'csv-file-creator'
+import update     from "immutability-helper"
 //import Switch from 'react-toggle'
 
 //import bookings from '../bookings'
@@ -17,9 +18,11 @@ import {
     CardBody,
     CardTitle,
     Table
-} from 'reactstrap';
+}                 from 'reactstrap';
 import ageFactory from "../../age";
+import feeFactory from "../../../shared/fee/feeFactory";
 
+Moment.locale('en-gb');
 
 //import W from '../../../shared/woodcraft.js'
 
@@ -55,8 +58,22 @@ export default class Bookings extends React.Component {
     }
 
     exportCSV() {
-        const exportedData = this.props.bookings.map(b => [b.id,
+
+        const event = this.props.Event.toJS();
+
+        this.getFeesOwed = feeFactory(event).getFeesOwed;
+
+        const exportedData = this.props.bookings.map(b => {
+
+            let owed = this.getFeesOwed(event, b.participants, b).reduce((a, c) => parseFloat(c.total) + a, 0);
+            const paid = b.payments.filter(p => p.type === 'payment').reduce((a, c) => a + parseFloat(c.amount), 0);
+
+            owed = b.payments.filter(p => p.type === 'adjustment').reduce((a, c) => a + parseFloat(c.amount), owed);
+
+
+            return [b.id,
             b.userName,
+                b.district,
             b.userEmail,
             b.userContact,
             b.participants.length,
@@ -64,10 +81,14 @@ export default class Bookings extends React.Component {
             b.emergencyName,
             b.emergencyPhone,
             b.note,
+                '£' + owed,
+                '£' + paid,
+                '£' + (owed - paid),
             b.createdAt,
-            b.updatedAt]);
+                b.updatedAt]
+        });
         const fileName = this.props.Event.get('name') + "-Bookings-" + Moment().format('YYYY-MM-DD') + ".csv";
-        csv(fileName, [['id', 'Name', 'e-mail', 'Phone', 'Participants', 'Payment type', 'Emergency name', 'Emergency Contact', 'Note', 'Created', 'Updated'], ...exportedData]);
+        csv(fileName, [['id', 'Name', 'District', 'e-mail', 'Phone', 'Participants', 'Payment type', 'Emergency name', 'Emergency Contact', 'Note', 'Money Owed', 'Money Paid', 'Outstanding Balance', 'Created', 'Updated'], ...exportedData]);
     }
 
     markPaid(id) {
@@ -78,7 +99,7 @@ export default class Bookings extends React.Component {
 
     render() {
 
-        //const event = this.props.Event.toJS();
+        const event = this.props.Event.toJS();
         const bookings = this.props.bookings;
         const user = this.props.User.toJS();
         //const participants = this.props.Participants.toJS();
@@ -93,14 +114,15 @@ export default class Bookings extends React.Component {
     */
         const data = bookings.map(b => {
             let result = {
-                userName: b.userName,
-                userEmail: b.userEmail,
+                district:    b.district,
+                userName:    b.userName,
+                userEmail:   b.userEmail,
                 userContact: b.userContact,
                 paymentType: b.paymentType,
-                paid: b.paid,
-                id: b.id,
-                b: b,
-                E: this.props.Event
+                paid:        b.paid,
+                id:          b.id,
+                b:           b,
+                E:           this.props.Event
 
             };
 
@@ -116,20 +138,39 @@ export default class Bookings extends React.Component {
         });
 
 
-        const columns = [{accessor: "userName", Header: "Name", sortable: true},
-            {accessor: "userEmail", Header: "e-mail", sortable: true},
-            {accessor: "userContact", Header: "Contact", sortable: true, minWidth: 70},
+        const columns = [];
+
+        if (event.bigCampMode) columns.push({accessor: "district", Header: "District", sortable: true});
+
+        columns.push(...[{accessor: "userName", Header: "Name", sortable: true},
+            {
+                accessor: "userEmail", Header: "e-mail", sortable: true,
+                Cell:     row => <a href={`mailto:${row.original.userEmail}`}
+                                    target="_blank">{row.original.userEmail}</a>
+            },
+            {
+                accessor: "userContact", Header: "Contact", sortable: true, minWidth: 70,
+                Cell:     row => <a href={`tel:${row.original.userContact}`}
+                                    target="_blank">{row.original.userContact}</a>
+            },
             {accessor: "participants", Header: "Booked", sortable: true, minWidth: 50},
             {accessor: "paymentType", Header: "Payment Method", sortable: true, minWidth: 50},
-            {accessor: "updatedAt", Header: "Updated", sortable: true, minWidth: 40},
+            {id:            'updatedAt',
+                accessor:   row => row,
+                Cell:       row => row.original.updatedAt,
+                sortMethod: updatedSort,
+                Header:     "Updated",
+                sortable:   true,
+                minWidth:   40
+            },
             {
-                id: "edit",
+                id:       "edit",
                 accessor: "id",
-                Header: "Edit",
-                Cell: row => <BookingEditLink event={row.original.E.toJS()} booking={row.original.b}/>,
+                Header:   "Edit",
+                Cell:     row => <BookingEditLink event={row.original.E.toJS()} booking={row.original.b}/>,
                 minWidth: 18
             }
-        ];
+        ]);
         //const sortables=[{column:"userName", sortFunction:nameSort},"userEmail","userContact","participants","paymentType","updatedAt"];
         const expanded = {[this.state.expanded]: true};
 
@@ -211,6 +252,9 @@ export default class Bookings extends React.Component {
     }
 }
 
+const updatedSort = (a, b) => {
+    return a.b.updatedAt > b.b.updatedAt ? 1 : -1;
+};
 
 
 const nameSort = (a, b) => {
