@@ -4,9 +4,10 @@ require("@babel/register");
 
 require("../config.js")()//config returns a promise the first time then overwrites its own module.exports to return a plain object on subsequent requires.
     .then(config => {
+        const log = require("./logging.js");
+
         const db = require('./orm.js');
 
-        const log = require("./logging.js");
         const backup = require("./backup");
 
         const express = require('express');
@@ -67,68 +68,29 @@ require("../config.js")()//config returns a promise the first time then overwrit
         server.get('/auth/facebook',
             passport.authenticate('facebook', {scope: ['email']}));
 
-        server.get('/auth/facebook/callback',
-            passport.authenticate('facebook', {failureRedirect: '/user'}),
-            function (req, res) {
-                // Successful authentication, redirect home.
-                log.info({
-                    message: "User logged in from Facebook {ip} {session} {user} {email}",
-                    ip: req.ip,
-                    session: req.session.id,
-                    user: req.user.userName,
-                    email: req.user.email
-                });
-                res.redirect('/');
-            });
+        server.get('/auth/facebook/callback', function(req, res, next) {
+            passport.authenticate('facebook', authenticateCallback('Facebook', req, res, next))(req, res, next);
+        });
 
         server.get('/auth/google',
             passport.authenticate('google', {scope: ['email', 'profile']})); //google OAuth redirect
 
-        server.get('/auth/google/callback',
-            passport.authenticate('google', {failureRedirect: '/user'}),
-            function (req, res) {
-                // Successful authentication, redirect home.
-                log.info({
-                    message: "User logged in from Google {ip} {session} {user} {email}",
-                    ip: req.ip,
-                    session: req.session.id,
-                    user: req.user.userName,
-                    email: req.user.email
-                });
-                res.redirect('/');
-            }); // google OAuth callback
+        server.get('/auth/google/callback', function(req, res, next) {
+            passport.authenticate('google', authenticateCallback('Google', req, res, next))(req, res, next);
+        });
 
         server.get('/auth/yahoo',
             passport.authenticate('yahoo')); //google OAuth redirect
 
-        server.get('/auth/yahoo/callback',
-            passport.authenticate('yahoo', {failureRedirect: '/user'}),
-            function (req, res) {
-                // Successful authentication, redirect home.
-                log.info({
-                    message: "User logged in from Yahoo {ip} {session} {user} {email}",
-                    ip: req.ip,
-                    session: req.session.id,
-                    user: req.user.userName,
-                    email: req.user.email
-                });
-                res.redirect('/');
-            }); // google OAuth callback
+        server.get('/auth/yahoo/callback', function(req, res, next) {
+            passport.authenticate('yahoo', authenticateCallback('Yahoo', req, res, next))(req, res, next);
+        });
 
         server.get('/auth/microsoft', passport.authenticate('azuread-openidconnect', {failureRedirect: '/'}));
 
-        server.post('/auth/microsoft/callback',
-            passport.authenticate('azuread-openidconnect', {failureRedirect: '/'}),
-            (req, res) => {
-                log.info({
-                    message: "User logged in from Microsoft {ip} {session} {user} {email}",
-                    ip: req.ip,
-                    session: req.session.id,
-                    user: req.user.userName,
-                    email: req.user.email
-                });
-                res.redirect('/');
-            });
+        server.post('/auth/microsoft/callback', function(req, res, next) {
+            passport.authenticate('azuread-openidconnect', authenticateCallback('Microsoft', req, res, next))(req, res, next);
+        });
 
         server.get('/api/env', (req, res) => res.json({env: config.ENV}));
 
@@ -167,6 +129,9 @@ require("../config.js")()//config returns a promise the first time then overwrit
 
         server.post('/api/payment/add', P.addPayment, bookings.addPayment);
         server.post('/api/payment/delete', P.addPayment, bookings.deletePayment);
+
+        server.post('/api/membership/approve', P.updateMembership, bookings.approveMembership);
+        server.post('/api/membership/unapprove', P.updateMembership, bookings.unapproveMembership);
 
         if (config.ENV === 'dev') {
 
@@ -260,6 +225,42 @@ require("../config.js")()//config returns a promise the first time then overwrit
                 });
             }
             next();
+        }
+
+        function authenticateCallback(source, req, res, next) {
+            return function (err, user, info) {
+                if (err && err.constructor.name === 'WrongProviderError') {
+                    log.info({
+                        message: "Wrong provider login {ip} {session} used {original} expected {used}",
+                        ip: req.ip,
+                        session: req.session.id,
+                        user: req.user.userName,
+                        original:err.original,
+                        used: err.used
+                    });
+                    return res.redirect('/user/' + err.original)
+                }
+
+                if (err) {
+                    return next(err);
+                }
+                if (!user) {
+                    return res.redirect('/login');
+                }
+                log.info({
+                    message: `User logged in from ${source} {ip} {session} {user} {email}`,
+                    ip: req.ip,
+                    session: req.session.id,
+                    user: user.userName,
+                    email: user.email
+                });
+                req.logIn(user, function (err) {
+                    if (err) {
+                        return next(err);
+                    }
+                    return res.redirect('/');
+                });
+            }
         }
 
     }).catch(e => {
