@@ -16,6 +16,8 @@ import {
     Box
 } from 'react-html-email'
 
+import update from 'immutability-helper';
+
 import moment from 'moment';
 
 
@@ -29,20 +31,34 @@ export class Config extends React.Component {
     constructor(props) {
         super(props);
 
-        this.update = this.update.bind(this);
+        this.updateStandardFee = this.updateStandardFee.bind(this);
+        this.updateDiscountFee = this.updateDiscountFee.bind(this);
+        this.updateUnaccompaniedFee = this.updateUnaccompaniedFee.bind(this);
+        this.updateUnaccompaniedDiscountFee = this.updateUnaccompaniedDiscountFee.bind(this);
     }
 
-    update(e) {
-        const fee = {amount: parseFloat(e.target.value)}
-        this.props.update(fee);
+    updateStandardFee(e) {
+        this.props.update(update(this.props.fee, {amount:{$set:  parseFloat(e.target.value)}}));
+    }
+
+    updateDiscountFee(e) {
+        this.props.update(update(this.props.fee, {amountDiscount:{$set:  parseFloat(e.target.value)}}));
+    }
+
+    updateUnaccompaniedFee(e) {
+        this.props.update(update(this.props.fee, {amountUnaccompanied:{$set:  parseFloat(e.target.value)}}));
+    }
+
+    updateUnaccompaniedDiscountFee(e) {
+        this.props.update(update(this.props.fee, {amountUnaccompaniedDiscount:{$set:  parseFloat(e.target.value)}}));
     }
 
     render() {
         //Thou shalt not ever use JS numbers for currency...
         const amount = this.props.fee.amount || 35;
-        const unaccompanied = amount === 35 ? 50 : amount * 1.5;
-        const unaccompaniedDiscount = amount === 35 ? 25 : amount * 0.75;
-        const discount = amount === 35 ? 20 : amount * 0.5;
+        const unaccompanied = this.props.fee.amountUnaccompanied || 50;
+        const unaccompaniedDiscount = this.props.fee.amountUnaccompaniedDiscount || 25;
+        const discount = this.props.fee.amountDiscount || 17;
 
         return (<Row>
             <Col>
@@ -57,8 +73,24 @@ export class Config extends React.Component {
                     <tbody>
                     <tr>
                         <td>Unaccompanied Elfins, Pioneers &amp; Venturers</td>
-                        <td>£{Math.round(unaccompanied)}</td>
-                        <td>£{Math.round(unaccompaniedDiscount)}</td>
+                        <td><FormGroup>
+                            <InputGroup>
+                                <div className="input-group-prepend">
+                                    <span className="input-group-text">£</span>
+                                </div>
+                                <Input type="number" className="form-control" placeholder="50" value={unaccompanied}
+                                       onChange={this.updateUnaccompaniedFee}/>
+                            </InputGroup>
+                        </FormGroup></td>
+                        <td><FormGroup>
+                            <InputGroup>
+                                <div className="input-group-prepend">
+                                    <span className="input-group-text">£</span>
+                                </div>
+                                <Input type="number" className="form-control" placeholder="25" value={unaccompaniedDiscount}
+                                       onChange={this.updateUnaccompaniedDiscountFee}/>
+                            </InputGroup>
+                        </FormGroup></td>
                     </tr>
                     <tr>
                         <td>Elfins, Pioneers &amp; Venturers accompanied by a responsible adult, DFs and Adults</td>
@@ -69,11 +101,19 @@ export class Config extends React.Component {
                                         <span className="input-group-text">£</span>
                                     </div>
                                     <Input type="number" className="form-control" placeholder="35" value={amount}
-                                           onChange={this.update}/>
+                                           onChange={this.updateStandardFee}/>
                                 </InputGroup>
                             </FormGroup>
                         </td>
-                        <td>£{Math.round(discount)}</td>
+                        <td><FormGroup>
+                            <InputGroup>
+                                <div className="input-group-prepend">
+                                    <span className="input-group-text">£</span>
+                                </div>
+                                <Input type="number" className="form-control" placeholder="17" value={discount}
+                                       onChange={this.updateDiscountFee}/>
+                            </InputGroup>
+                        </FormGroup></td>
                     </tr>
                     </tbody>
                 </Table>
@@ -88,13 +128,13 @@ export class BookingForm extends React.Component {
 
         const accompanied = this.props.participants.find(p => moment(this.props.event.startDate).diff(moment(p.age), 'years') > 15) === undefined ? false : true;
 
-        const amount = this.props.feeData.amount;
-        const unaccompanied = amount === 35 ? 50 : amount * 1.5;
-        const unaccompaniedDiscount = amount === 35 ? 25 : amount * 0.75;
-        const discount = amount === 35 ? 20 : amount * 0.5;
+        const amount = this.props.feeData.amount || 35;
+        const unaccompanied = this.props.feeData.amountUnaccompanied || 50;
+        const unaccompaniedDiscount = this.props.feeData.amountUnaccompaniedDiscount || 25;
+        const discount = this.props.feeData.amountDiscount || 17;
 
-        const total = this.props.participants.length * Math.round(accompanied ? amount : unaccompanied);
-        const totalDiscounted = this.props.participants.length * Math.round(accompanied ? discount : unaccompaniedDiscount);
+        const total = this.props.participants.length * (accompanied ? amount : unaccompanied);
+        const totalDiscounted = this.props.participants.length * (accompanied ? discount : unaccompaniedDiscount);
 
         return (<div className="col-sm-12">
             <table className="table">
@@ -108,20 +148,20 @@ export class BookingForm extends React.Component {
                 <tbody>
                 <tr>
                     <td>Unaccompanied Elfins, Pioneers &amp; Venturers</td>
-                    <td>£{Math.round(unaccompanied)}</td>
-                    <td>£{Math.round(unaccompaniedDiscount)}</td>
+                    <td>£{unaccompanied.toFixed(2)}</td>
+                    <td>£{unaccompaniedDiscount.toFixed(2)}</td>
                 </tr>
                 <tr>
                     <td>Elfins, Pioneers &amp; Venturers accompanied by a responsible adult, DFs and Adults</td>
-                    <td>£{Math.round(amount)}</td>
-                    <td>£{Math.round(discount)}</td>
+                    <td>£{amount.toFixed(2)}</td>
+                    <td>£{discount.toFixed(2)}</td>
                 </tr>
                 <tr>
                     <td><b>My
                         Booking</b> ({this.props.participants.length} {this.props.participants.length < 2 ? "person" : "people"}, {accompanied ? "Accompanied" : "Unaccompanied"})
                     </td>
-                    <td><b>£{total}</b></td>
-                    <td><b>£{totalDiscounted}</b></td>
+                    <td><b>£{total.toFixed(2)}</b></td>
+                    <td><b>£{totalDiscounted.toFixed(2)}</b></td>
                 </tr>
                 </tbody>
             </table>
@@ -142,13 +182,13 @@ export class ThanksRow extends React.Component {
 
         const accompanied = this.props.booking.participants.find(p => moment(this.props.event.startDate).diff(moment(p.age), 'years') > 15) === undefined ? false : true;
 
-        const amount = this.props.event.feeData.amount;
-        const unaccompanied = amount === 35 ? 50 : amount * 1.5;
-        const unaccompaniedDiscount = amount === 35 ? 25 : amount * 0.75;
-        const discount = amount === 35 ? 20 : amount * 0.5;
+        const amount = this.props.event.feeData.amount || 35;
+        const unaccompanied = this.props.event.feeData.amountUnaccompanied || 50;
+        const unaccompaniedDiscount = this.props.event.feeData.amountUnaccompaniedDiscount || 25;
+        const discount = this.props.event.feeData.amountDiscount || 17;
 
-        const total = this.props.booking.participants.length * Math.round(accompanied ? amount : unaccompanied);
-        const totalDiscounted = this.props.booking.participants.length * Math.round(accompanied ? discount : unaccompaniedDiscount);
+        const total = this.props.booking.participants.length * (accompanied ? amount : unaccompanied);
+        const totalDiscounted = this.props.booking.participants.length * (accompanied ? discount : unaccompaniedDiscount);
 
         return (<Row>
             <Col>
@@ -164,20 +204,20 @@ export class ThanksRow extends React.Component {
                     <tbody>
                     <tr>
                         <td>Unaccompanied Elfins, Pioneers &amp; Venturers</td>
-                        <td>£{Math.round(unaccompanied)}</td>
-                        <td>£{Math.round(unaccompaniedDiscount)}</td>
+                        <td>£{unaccompanied.toFixed(2)}</td>
+                        <td>£{unaccompaniedDiscount.toFixed(2)}</td>
                     </tr>
                     <tr>
                         <td>Elfins, Pioneers &amp; Venturers accompanied by a responsible adult, DFs and Adults</td>
-                        <td>£{Math.round(amount)}</td>
-                        <td>£{Math.round(discount)}</td>
+                        <td>£{amount.toFixed(2)}</td>
+                        <td>£{discount.toFixed(2)}</td>
                     </tr>
                     <tr>
                         <td><b>My
                             Booking</b> ({this.props.booking.participants.length} {this.props.booking.participants.length < 2 ? "person" : "people"}, {accompanied ? "Accompanied" : "Unaccompanied"})
                         </td>
-                        <td><b>£{total}</b></td>
-                        <td><b>£{totalDiscounted}</b></td>
+                        <td><b>£{total.toFixed(2)}</b></td>
+                        <td><b>£{totalDiscounted.toFixed(2)}</b></td>
                     </tr>
                     </tbody>
                 </Table>
@@ -208,15 +248,15 @@ export function emailHTML(event, booking) {
 export function getFeesOwed(event, participants, booking) {
     const accompanied = participants.find(p => moment(event.startDate).diff(moment(p.age), 'years') > 15) === undefined ? false : true;
 
-    const amount = event.feeData.amount;
-    const unaccompanied = amount === 35 ? 50 : amount * 1.5;
-    const unaccompaniedDiscount = amount === 35 ? 25 : amount * 0.75;
-    const discount = amount === 35 ? 20 : amount * 0.5;
+    const amount = event.feeData.amount || 35;
+    const unaccompanied = event.feeData.amountUnaccompanied || 50;
+    const unaccompaniedDiscount = event.feeData.amountUnaccompaniedDiscount || 25;
+    const discount = event.feeData.amountDiscount || 17;
 
-    const total = participants.length * Math.round(accompanied ? amount : unaccompanied);
-    const totalDiscounted = participants.length * Math.round(accompanied ? discount : unaccompaniedDiscount);
+    const total = participants.length * (accompanied ? amount : unaccompanied);
+    const totalDiscounted = participants.length * (accompanied ? discount : unaccompaniedDiscount);
 
     const line = `${participants.length} ${participants.length < 2 ? "person" : "people"}, ${accompanied ? "Accompanied" : "Unaccompanied"} (Discount: £${totalDiscounted})`;
 
-    return [{line: line, total: total}];
+    return [{line: line, total: total.toFixed(2)}];
 }
